refactor: Options move `Tdd\Runner` => `Tdd\Command`
